\name{coord}
\alias{coord}
\alias{example_coord}

\title{Input file for \code{\link{TESS3}}}
\description{
Description of the \code{coord} format.
The \code{coord} format can be used as an input format 
for coordinates matrices in the functions \code{\link{TESS3}}. 
}
\details{
The \code{geno} format has one row for each individual. Each row contains the longitude and latitude
coordinates of each individual.

Here is an example of a coordiantes matrix using the \code{coord} format with 3 
individuals and 4 loci: 
\preformatted{
2.5154 5.4390
-8.4293 4.0197
1.3536 5.5852
}
}

\author{
    Kevin Caye
}
\seealso{

}


\keyword{format}
