% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3utils.R
\name{createGridFromAsciiRaster}
\alias{createGridFromAsciiRaster}
\title{Create grid on which coefficients will be displayed}
\usage{
createGridFromAsciiRaster(file)
}
\arguments{
\item{file}{A character string containing a the path to the ascii raster file.}
}
\value{
\code{createGridFromAsciiRaster} returns a matrix which represent the grid extract from the raster file.
}
\description{
\code{createGridFromAsciiRaster} returns a grid to use to project clusters on a map. The grid is computed from an ascii raster file.
}
\seealso{
This function was made to be used with \code{\link{getConstraintsFromAsciiRaster}} and \code{\link{maps}}.
}

