% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tess3utils.R
\name{getConstraintsFromAsciiRaster}
\alias{getConstraintsFromAsciiRaster}
\title{Create grid on which coefficients will be displayed}
\usage{
getConstraintsFromAsciiRaster(file, cell_value_min = NULL,
  cell_value_max = NULL)
}
\arguments{
\item{file}{A character string containing a the path to the ascii raster file.}

\item{cell_value_min}{All case values inferior to this numeric value will be set to FALSE on the grid.}

\item{cell_value_max}{All case values superior to this numeric value will be set to FALSE on the grid.}
}
\value{
\code{getConstraintsFromAsciiRaster} returns a matrix with boolean cells.
Cells is set to TRUE if user wants to project on this map's cell
}
\description{
\code{getConstraintsFromAsciiRaster} compute a constraint matrix acording to the \code{cell_value_min} and \code{cell_value_max} values.
}
\seealso{
This function was made to be used with \code{\link{createGridFromAsciiRaster}} and \code{\link{maps}}.
}

