% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tess3utils.R
\name{maps}
\alias{maps}
\title{Display maps of membership/admixture coefficients.}
\usage{
maps(matrix, coord, grid, constraints = NULL, method = "treshold",
  colorGradientsList = lColorGradients, onemap = T, onepage = T, ...)
}
\arguments{
\item{matrix}{Matrix of values to interpolate.}

\item{coord}{Matrix of corrdinate values.}

\item{grid}{Matrix with grid values. Can be generated by \code{\link{createGridFromAsciiRaster}}.}

\item{constraints}{Matrix with constraints values. Can be generated by \code{\link{createGridFromAsciiRaster}}.}

\item{method}{A character string which can be "treshold" or "max".}

\item{colorGradientsList}{A list of string which defines what color are used to plot the interpolation result.}

\item{onemap}{A boolean value.}

\item{onepage}{A boolean value.}
}
\description{
\code{\link{maps}} projects membership/admixture coefficients on a grid using \code{\link[fields]{Krig}}.
Gradients in coefficients are represented by gradients in colors.
if \code{onemap == T & method == "treshold"} only \code{coefficients > 0.5} are plotted.
if \code{onemap == T & method == "max"} at each point the cluster for which the coefficient
is maximal is plotted (even if the value is less than 0.5).
if \code{onemap == F} all values are plotted (since there is one cluster represented on each map, there is no overlap problem).
}
\examples{
## How maps can be used to plot TESS3 function output ##

#################
# runs of TESS3 #
#################
# Retrieve data file name
genotype.file <- system.file("extdata/Athaliana","Athaliana.geno",package = "tess3r")
coord.file <- system.file("extdata/Athaliana","Athaliana.coord",package = "tess3r")
# Read coordinate file
coord <- read.coord(coord.file)
n <- nrow(coord)

project <- TESS3(input.file = genotype.file,
                 input.coord = coord.file,
                 K = 1:5,
                 ploidy = 1,
                 repetitions = 1,
                 entropy = TRUE,
                 percentage = 0.2,
                 project = "new")

#####################
# Plot result on map#
#####################

asciiFile=system.file("extdata/","lowResEurope.asc",package = "tess3r")
grid=createGridFromAsciiRaster(asciiFile)
# To display only altitudes above 0:
constraints=getConstraintsFromAsciiRaster(asciiFile,cell_value_min=0)

maps(matrix = Q( project, K = 3, run = 1 ),
     coord = coord,
     grid=grid,constraints=constraints,method="max",main="ancestry coefficient with K = 3")
}
\seealso{
This function was made to used with \code{\link{createGridFromAsciiRaster}}, \code{\link{maps}} and \code{\link{TESS3}}
}

