% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.dist.from.center.dirichlet.Q}
\alias{sample.dist.from.center.dirichlet.Q}
\title{sample pop center and compute dist.from.center as dirichlet parameter}
\usage{
sample.dist.from.center.dirichlet.Q()
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

