% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{sample.gaussian.process.Q}
\alias{sample.gaussian.process.Q}
\title{sample Q such Q = logit(a)
a ~ N(0,W) W_ij = cov.function(X_i,X_j)}
\usage{
sample.gaussian.process.Q(cov.function)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

