% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampleTESS2.3}
\alias{sampleTESS2.3}
\title{Sample X such that P(X_i_dl + j) = Sum(Q_ik G_kj).
Coord ...}
\usage{
sampleTESS2.3(n.by.pop, L, d, K,
  Q.sampler = sample.dist.from.center.dirichlet.Q())
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

