% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3utils.R
\name{tess2tess3}
\alias{tess2tess3}
\title{Convert allelic data in TESS 2.3 format into .geno format.}
\usage{
tess2tess3(file = "data.tess")
}
\description{
Convert allelic data in TESS 2.3 format into .geno format.
}

