% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tess3Project.R
\docType{class}
\name{tess3Project-class}
\alias{tess3Project-class}
\title{An S4 class to represent a TESS3 project.}
\description{
An S4 class to represent a TESS3 project.
}
\seealso{
You can see how to use method of this class in \code{\link{TESS3}}
}

