% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tess3.R
\docType{package}
\name{tess3r}
\alias{tess3r}
\alias{tess3r-package}
\title{tess3r : An R Package for Population Genetics Study}
\description{
This R package implements the TESS3 program and tools useful to plot program outputs.
}

